Add FSDP2 training and RoboTwin lazy rendering; finish multitask, remove   action-prefix
  Training:
  - FSDP2 support (per-parameter DTensor sharding via accelerate
  fsdp_version=2):
    FSDPConfig in train_config + train.py wiring (sharded checkpoint save/load,
    fp32 master weights, replicated-grad sync + mixed grad-norm clip helpers);
    add accelerate>=1.13.0. Modeling prerequisites: ModuleList layernorms,
    detach_backbone_layer_aliases, bf16 backbone_dtype, PI05ModelLayer.forward
    dispatcher; fp32 softmax in the attention SDPA path.
  - Rebuild the LR scheduler after the time_mlp_weight_decay param-group split
    (torch>=2.10 strict-zip crash on the first scheduler.step).

  RoboTwin:
  - Lazy rendering: ROBOTWIN_LAZY_RENDER skips the SAPIEN render on steps that
    replay a cached action; serve cached actions when observation is None; lazy
    obs poll + efficiency metrics in the eval client. Evaluate on the 'seen'
    instruction split.
  - Finish merged multitask: add roots / config_subdirs (pool clean+randomized
    per task) / stats_path (exact pooled-stats override) to
  RoboTwinMultiTaskConfig;
    make_robotwin_multitask_dataset builds one subset per (task, subdir); add
    pi05_flashvla_clean_merged_multitask.yaml.

  Cleanup / fixes:
  - Remove the dead use_action_prefix feature (always False) from
  pi05/pi0/smolvla
    config+modeling, the dataset, and the trainer (behavior-neutral).
  - compat.py: import load_json/cast_stats_to_numpy from
  lerobot.datasets.io_utils
    (not .utils, where they don't exist) — fixes ImportError on v2.1 stats
  loading.

+++ flashvla/layers/attention.py
@@ -121,16 +121,16 @@ class Attention(nn.Module):
 
         if return_attn_probs:
             # Memory-efficient backend doesn't return attention weights.
-            attn_scores = torch.matmul(q, k.transpose(-2, -1)) * self.scale
+            attn_scores = torch.matmul(q.float(), k.float().transpose(-2, -1)) * self.scale
             if attention_mask is not None:
-                attn_scores = attn_scores + attention_mask
-            attn_weights = torch.softmax(attn_scores, dim=-1)
+                attn_scores = attn_scores + attention_mask.float()
+            attn_weights = torch.softmax(attn_scores, dim=-1).to(q.dtype)
             out = torch.matmul(attn_weights, v)
             return out, attn_weights
 
-        attn_scores = torch.matmul(q, k.transpose(-2, -1)) * self.scale
+        attn_scores = torch.matmul(q.float(), k.float().transpose(-2, -1)) * self.scale
         if attention_mask is not None:
-            attn_scores = attn_scores + attention_mask
-        attn_weights = torch.softmax(attn_scores, dim=-1)
+            attn_scores = attn_scores + attention_mask.float()
+        attn_weights = torch.softmax(attn_scores, dim=-1).to(q.dtype)
         out = torch.matmul(attn_weights, v)
         return out
